requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
fastapi
fasttext
filelock
fsspec
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httpx
huggingface-hub
hyperframe
idna
langdetect
numpy
packaging
portalocker
protobuf
pybind11
pydantic
pydantic_core
PyPDF2
python-dotenv
PyYAML
qdrant-client
regex
requests
safetensors
setuptools
six
sniffio
starlette
tokenizers
tqdm
transformers
typing_extensions
urllib3
uvicorn
